Move `sort_coverage` from `utils.py` to ` coverage.py`.
