+++ gcovr/writer/txt.py
@@ -21,11 +21,10 @@ from typing import Iterable, Tuple
 
 from ..utils import (
     force_unix_separator,
-    sort_coverage,
     presentable_filename,
     open_text_for_writing,
 )
-from ..coverage import CovData, CoverageStat, FileCoverage
+from ..coverage import CovData, CoverageStat, FileCoverage, sort_coverage
 
 # Widths of the various columns
 COL_FILE_WIDTH = 40
